// cms-next-typesxript/src/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IArticleFields {
  /** Title */
  title: string;

  /** slug */
  slug: string;

  /** Description */
  description: string;

  /** Content */
  content?: Document | undefined;

  /** Buttons text */
  actions?: string | undefined;
}

export interface IArticle extends Entry<IArticleFields> {
  fields: {
    actions: string;
    content: document;
    description: string;
    slug: string;
    title: string;
  };
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "article";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IHomeFields {
  /** Заголовок */
  title?: string | undefined;

  /** Текст статті */
  description?: Document | undefined;

  /** Зображення статті */
  background?: Asset | undefined;
}

export interface IHome extends Entry<IHomeFields> {
  fields: {
    title: string;

    /** Текст статті */
    description: Document;

    /** Зображення статті */
    background: Asset;
  };
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "home";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE = "article" | "home";

export type IEntry = IArticle | IHome;

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";